+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## 0.9.0 - Unreleased
+- Added a new Goto command: Goto Require (Fixes [#28](https://github.com/pedrorgirardi/Pep/issues/28))
+- Added a new Goto command: Goto Import (Fixes [#62](https://github.com/pedrorgirardi/Pep/issues/62))
+- Goto Analysis Finding - Removed file path from Quick Panel Item
+- Added a new setting 'analysis_applicable_to' to set applicable analysis syntaxes
+
 ## 0.8.1 - 2022-07-21
 - Use KIND_ID_ prefix to support older versions of Sublime Text